Using the get() method of the superclass Message
This commit defines a get() method for the superclass Message
which instantiates and returns the right subtype from amongst
FriendRequest, Challenge and Note, by using information from the
column named 'type' in the Message table.

I think this is better design than calling get() on the subtypes.

Take a look, @johnyangsam, @makazone